1.5 部署打包脚本优化

+++ {{cookiecutter.project_name}}/deploy/deploy-project.py
@@ -7,6 +7,7 @@
 #
 
 # !/usr/bin/env python
+import configparser
 import datetime
 import logging
 
@@ -17,6 +18,8 @@ import sys
 import os
 from pick import pick
 
+from {{cookiecutter.app_name}}.commons.utils.config_util import ini2dict
+
 logger = logging.getLogger(__name__)
 
 
@@ -25,27 +28,6 @@ class Config:
 
     def __init__(self, env_file_name):
         self.env_file_name = env_file_name
-
-
-ss_{{cookiecutter.app_name}}_test_config_dict1 = {
-    'server_ip': '127.0.0.1',
-    'server_name': '{{cookiecutter.app_name}}test',
-    'username': 'admin',
-    'base_path': '/home/sysop/docker/{{cookiecutter.app_name}}/',
-    'env_file_name': 'flask{{cookiecutter.app_name}}testenv',
-    'config_base_path': 'ss-{{cookiecutter.app_name}}-test-01'
-}
-
-ss_{{cookiecutter.app_name}}_pro_config_dict1 = {
-    'server_ip': '127.0.0.1',
-    'server_name': 'admin',
-    'username': '{{cookiecutter.app_name}}ai',
-    'base_path': '/home/{{cookiecutter.app_name}}ai/docker/',
-    'env_file_name': 'flask{{cookiecutter.app_name}}proenv',
-    'config_base_path': '{{cookiecutter.project_name}}-server-01'
-}
-
-server_config = ss_{{cookiecutter.app_name}}_test_config_dict1
 
 
 def sftp_transfer_callback(x, y):
@@ -58,12 +40,12 @@ def sftp_transfer_callback(x, y):
     sys.stdout.flush()
 
 
-@retry(tries=4, delay=5)
-def sftp_upload_file(server_ip, local_path, server_path, is_dir=False):
-    with pysftp.Connection(host=server_ip, username=username,
+# @retry(tries=4, delay=5)
+def sftp_upload_file(ip, path, server_path, is_dir=False):
+    with pysftp.Connection(host=ip, username=username,
                            private_key=os.path.expanduser('~/.ssh/id_rsa')) as sftp:
         print("Connection succesfully stablished ... ")
-        print(local_path)
+        print(path)
         print(server_path)
         if is_dir:
             sftp.makedirs(server_path)
@@ -71,9 +53,9 @@ def sftp_upload_file(server_ip, local_path, server_path, is_dir=False):
             parent_path = os.path.dirname(server_path)
             sftp.makedirs(parent_path)
         if is_dir:
-            sftp.put_r(localpath=local_path, remotepath=server_path)
+            sftp.put_r(localpath=path, remotepath=server_path)
         else:
-            sftp.put(localpath=local_path, remotepath=server_path,
+            sftp.put(localpath=path, remotepath=server_path,
                      callback=sftp_transfer_callback)
 
 
@@ -116,16 +98,13 @@ def get_docker_entry_point_full_path(config_base_path, entry_point_file_name):
 config_list = list()
 
 
-def build_and_package():
-    pyarmor_cmd = 'pyarmor obfuscate --exclude ' \
-                  + algo_exclude_path + ' --recursive --output ' + algo_bin_path + ' ' \
-                  + algo_src_path + '/__init__.py --platform linux64'
-    print(pyarmor_cmd)
-    os.system(pyarmor_cmd)
-    ## 3 编译前端
+def build_and_package_fe():
+    # 编译打包前端
     fe_npm_cmd = 'cd ' + fe_src_path
-    fe_npm_cmd +=';npm run build:stage' if fe_build_test else ';npm run build:prod'
-    fe_npm_cmd +='; gitbook build src/docs dist/docs'
+    # fe_npm_cmd += ';npm run build:stage' if fe_build_test else ';npm run build:prod'
+    fe_npm_cmd += ';npm run build'
+    # 如果有 文档编译文档
+    # fe_npm_cmd += '; gitbook build src/docs dist/docs'
     print(fe_npm_cmd)
     os.system(fe_npm_cmd)
     fe_mv_cmd = 'mv ' + fe_src_path + '/dist/* ' + fe_bin_path
@@ -134,7 +113,15 @@ def build_and_package():
     fe_tar_cmd = 'cd ' + fe_bin_path + '; tar -zcvf ../../{{cookiecutter.app_name}}-fe.tar.gz static index.html favicon.ico docs'
     print(fe_tar_cmd)
     os.system(fe_tar_cmd)
-    # 4 打包后端
+
+
+def build_and_package_be():
+    # 编译打包后端代码
+    # 1. 加密核心算法
+    pyarmor_cmd = f'pyarmor obfuscate --exclude {algo_exclude_path } --recursive --output {algo_bin_path} {algo_src_path }/__init__.py --platform linux64'
+    print(pyarmor_cmd)
+    os.system(pyarmor_cmd)
+    # 2 打包后端
     dirs = ['{{cookiecutter.app_name}}']
     files = [
         'setup.py',
@@ -146,14 +133,15 @@ def build_and_package():
     all_files.extend(files)
     print(all_files)
     for dir in dirs:
-        os.system('cp -r ' + os.path.join(be_server_src_path, dir) + ' ' + bin_path)
+        os.system('cp -r ' + os.path.join(be_server_src_path, dir) + ' ' + be_bin_path)
     for f in files:
-        os.system('cp ' + os.path.join(be_server_src_path, f) + ' ' + bin_path)
-    print(bin_path)
-    all_files.append('model')
-    # all_files.append('fe')
+        os.system('cp ' + os.path.join(be_server_src_path, f) + ' ' + be_bin_path)
+    cp_cmd = 'cp ' + get_env_file_full_path(config_base_path, 'logging.yml') + ' ' + be_bin_path
+    print(cp_cmd)
+    os.system(cp_cmd)
+    print(be_bin_path)
     tar_filenames = ' '.join(all_files)
-    tar_cmd = 'cd ' + bin_path + '; tar -zcvf ../{{cookiecutter.app_name}}.tar.gz ' + tar_filenames
+    tar_cmd = 'cd ' + be_bin_path + '; tar -zcvf ../../{{cookiecutter.app_name}}-be.tar.gz ' + tar_filenames
     print(tar_cmd)
     os.system(tar_cmd)
 
@@ -164,8 +152,8 @@ def upload():
     sftp_upload_file(server_ip, os.path.join(project_dir, 'deploy/Dockerfile'), remote_path + 'Dockerfile')
     sftp_upload_file(server_ip, get_env_file_full_path(config_base_path, env_file_name), remote_path + env_file_name)
     sftp_upload_file(server_ip, get_base_image_full_path(), remote_path + '{{cookiecutter.project_name}}_base_image', True)
-    sftp_upload_file(server_ip, be_zip_file_path, remote_path + '{{cookiecutter.app_name}}.tar.gz')
-    sftp_upload_file(server_ip, fe_zip_file_path, remote_path + 'nginx/{{cookiecutter.app_name}}-fe.tar.gz')
+    sftp_upload_file(server_ip, be_zip_file_path, remote_path + '{{cookiecutter.app_name}}-be.tar.gz')
+    sftp_upload_file(server_ip, fe_zip_file_path, remote_path + '{{cookiecutter.app_name}}-fe.tar.gz')
 
 
 def remote_build():
@@ -200,21 +188,38 @@ def remote_build():
     pass
 
 
-def clean_all():
-    os.system('rm -rf ' + dist_path + '/*')
+def confirm_selector(title):
+    options = [{'label': '是', 'value': True}, {'label': '否', 'value': False}]
+    selected, idx = pick(options,
+                         title,
+                         multi_select=False,
+                         min_selection_count=1,
+                         options_map_func=lambda option: option.get('label'))
+    return selected.get('value')
+
+
+def clean_all(force=True):
+    if not force:
+        rst = confirm_selector('是否删除dist目录 ' + dist_path)
+        if rst:
+            os.system('rm -rf ' + dist_path + '/*')
+    else:
+        os.system('rm -rf ' + dist_path + '/*')
 
 
 def clean_other():
-    os.system('rm -rf ' + bin_path + '/*')
-    os.system('rm -f ' + be_zip_file_path)
-    print(be_zip_file_path)
-    os.system('rm -f ' + fe_zip_file_path)
-    print(fe_zip_file_path)
-    os.system('rm -rf ' + {{cookiecutter.app_name}}_package_dic_path)
-    print({{cookiecutter.app_name}}_package_dic_path)
+    rst = confirm_selector('是否删除dist目录 ' + dist_path)
+    if rst:
+        os.system('rm -rf ' + root_bin_path + '/*')
+        os.system('rm -f ' + be_zip_file_path)
+        print(be_zip_file_path)
+        os.system('rm -f ' + fe_zip_file_path)
+        print(fe_zip_file_path)
+        os.system('rm -rf ' + {{cookiecutter.app_name}}_package_dic_path)
+        print({{cookiecutter.app_name}}_package_dic_path)
 
 
-def package_script():
+def package_files():
     os.system('mkdir -p ' + os.path.join({{cookiecutter.app_name}}_package_dic_path, 'nginx'))
     os.system('mkdir -p ' + os.path.join({{cookiecutter.app_name}}_package_dic_path, 'sql'))
     cp_cmd = 'cp -R ' + be_zip_file_path + ' ' + {{cookiecutter.app_name}}_package_dic_path
@@ -231,13 +236,13 @@ def package_script():
     os.system(cp_cmd)
     cp_cmd = 'cp -R ' + get_env_file_full_path(config_base_path, 'deploy.sh') + ' ' + {{cookiecutter.app_name}}_package_dic_path
     os.system(cp_cmd)
-    cp_cmd = 'cp -R ' + get_base_image_full_path() + ' ' + os.path.join({{cookiecutter.app_name}}_package_dic_path, '{{cookiecutter.project_name}}_base_image')
+    cp_cmd = 'cp -R ' + get_base_image_full_path() + ' ' + os.path.join({{cookiecutter.app_name}}_package_dic_path, '{{cookiecutter.app_name}}_base_image')
     os.system(cp_cmd)
     cp_cmd = 'cp -R ' + os.path.join(project_dir, 'sql/{{cookiecutter.app_name}}.sql') + ' ' + os.path.join({{cookiecutter.app_name}}_package_dic_path, 'sql')
     os.system(cp_cmd)
     cp_cmd = 'cp -R ' + os.path.join(project_dir, 'sql/init-data.sql') + ' ' + os.path.join({{cookiecutter.app_name}}_package_dic_path, 'sql')
     os.system(cp_cmd)
-    tar_cmd = 'cd ' + dist_path + '; tar -zcvf ' + server_config['server_name'] + '-package' + date_str + '.tar.gz ' + server_config['server_name'] + '/*'
+    tar_cmd = 'cd ' + dist_path + '; tar -zcvf ' + server_config['section_name'] + '-package' + date_str + '.tar.gz ' + server_config['section_name'] + '/*'
     print(tar_cmd)
     os.system(tar_cmd)
     # os.system('mkdir -p ' + bin_path)
@@ -246,9 +251,12 @@ def package_script():
 if __name__ == '__main__':
     print("开始时间: " + datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
     try:
+        deploy_config = configparser.ConfigParser()
+        deploy_config.read('deploy_config.ini')
+        env_options = ini2dict(deploy_config)
         title1 = '请选择环境:'
-        option1 = ['vk-vas-test-01(172.16.206.60)', 'vk-vas-pro-02(127.0.0.1)']
-        selected1, idx1 = pick(option1, title1, multi_select=False, min_selection_count=1)
+        option1 = env_options
+        selected1, idx1 = pick(option1, title1, multi_select=False, min_selection_count=1, options_map_func=lambda option: f'{option.get("section_name")}({option.get("server_ip")})')
         title2 = '请选择前端打包环境:'
         option2 = ['测试', '生产']
         selected2, idx2 = pick(option2, title2, multi_select=False, min_selection_count=1)
@@ -264,11 +272,7 @@ if __name__ == '__main__':
             selected4, idx4 = pick(option4, title4, multi_select=False, min_selection_count=1)
             if idx4 == 1:
                 rebuild_base_image = True
-        server_config = None
-        if idx1 == 0:
-            server_config = ss_{{cookiecutter.app_name}}_test_config_dict1
-        elif idx1 == 1:
-            server_config = ss_{{cookiecutter.app_name}}_pro_config_dict1
+        server_config = selected1
         if server_config is None:
             raise Exception('select index error')
         print(os.getcwd())
@@ -279,37 +283,44 @@ if __name__ == '__main__':
         username = server_config['username']
         base_path = server_config['base_path']
         env_file_name = server_config['env_file_name']
-        config_base_path = server_config['config_base_path']
+        config_base_path = server_config['section_name']
         print('config_full_base_path==>' + get_config_full_base_path(config_base_path))
         print('nginx_config_full_base_path==>' + get_nginx_config_full_base_path(config_base_path))
-        algo_src_path = os.path.abspath(os.path.join(os.getcwd(), "../../", '{{cookiecutter.app_name}}_algo/model'))
+        # 算法源码目录
+        algo_src_path = os.path.abspath(os.path.join(os.getcwd(), server_config['algo_path']))
         be_server_src_path = os.path.abspath(os.path.join(os.getcwd(), '../'))
-        fe_src_path = os.path.abspath(os.path.join(os.getcwd(), "../../../", '{{cookiecutter.app_name}}-fe'))
-        algo_exclude_path = os.path.abspath(os.path.join(algo_src_path, 'demo'))
+        fe_src_path = os.path.abspath(os.path.join(os.getcwd(), server_config['fe_path']))
+        # 算法目录中不需要的么了
+        algo_exclude_path = os.path.abspath(os.path.join(algo_src_path, 'test'))
         dist_path = os.path.join(os.getcwd(), 'dist')
+        # 清除上次文件
         clean_all()
-        bin_path = os.path.join(dist_path, '{{cookiecutter.app_name}}')
-        os.system('mkdir -p ' + bin_path)
-        algo_bin_path = os.path.join(bin_path, 'model')
-        be_server_bin_path = os.path.join(bin_path, 'be-server')
-        fe_bin_path = os.path.join(bin_path, 'fe')
+
+        root_bin_path = os.path.join(dist_path, '{{cookiecutter.app_name}}')
+        be_bin_path = os.path.join(root_bin_path, '{{cookiecutter.app_name}}-be')
+        algo_bin_path = os.path.join(be_bin_path, '{{cookiecutter.app_name}}')
+        os.system('mkdir -p ' + be_bin_path)
+        fe_bin_path = os.path.join(root_bin_path, '{{cookiecutter.app_name}}-fe')
         os.system('mkdir -p ' + fe_bin_path)
         project_dir = os.path.abspath(os.path.join(os.getcwd(), "../"))
-        be_zip_file_path = dist_path + '/{{cookiecutter.app_name}}.tar.gz'
+        be_zip_file_path = dist_path + '/{{cookiecutter.app_name}}-be.tar.gz'
         fe_zip_file_path = dist_path + '/{{cookiecutter.app_name}}-fe.tar.gz'
-        {{cookiecutter.app_name}}_package_dic_path = os.path.join(dist_path, server_config['server_name'])
+        {{cookiecutter.app_name}}_package_dic_path = os.path.join(dist_path, server_config['section_name'])
         date_str = datetime.datetime.now().strftime('%Y%m%d')
-        {{cookiecutter.app_name}}_zip_file_path = dist_path + server_config['server_name'] + '-package' + date_str + '.tar.gz '
+        {{cookiecutter.app_name}}_zip_file_path = dist_path + server_config['section_name'] + '-package' + date_str + '.tar.gz '
         local_path = be_zip_file_path
         remote_path = base_path
         print(be_zip_file_path)
-        build_and_package()
-        if idx3 == 0:
+        # 打包核心代码
+        build_and_package_fe()
+        build_and_package_be()
+        if idx3 == 1:
             upload()
             remote_build()
-            clean_all()
+            clean_all(False)
         else:
-            package_script()
+            # 打包所有文件 包括配置
+            package_files()
             clean_other()
 
     except BaseException as e:
